refactor curated-recommendations to task in locustfile.py
