+++ tests/load/Dockerfile
@@ -41,7 +41,4 @@ COPY ./tests/load/locustfiles ./tests/load/locustfiles
 EXPOSE 8089 5557
 
 USER locust
-ENTRYPOINT ["locust", "-f", \
-  "tests/load/locustfiles/locustfile.py,\
-  tests/load/locustfiles/load.py,\
-  tests/load/locustfiles/curated_recommendations.py"]
+ENTRYPOINT ["locust", "-f", "tests/load/locustfiles/locustfile.py,tests/load/locustfiles/load.py"]
